requirements.txt
flask==2.3.3
flask-cors==3.0.10
easyocr==1.6.2
opencv-python-headless==4.8.1.78
numpy==1.25.2
Pillow==10.0.1
deep-translator==1.11.1
